requirements.txt
langchain
ipykernel
python-dotenv


langchain_community
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia

langchain-text-splitters

langchain-openai
chromadb

langchain-ollama

sentence_transformers
langchain_huggingface

faiss-cpu
# chromadb
langchain_chroma

langchain-openai
beautifulsoup4
streamlit